+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+- Activation will not choke on duplicate journeys/calls in route set.
+
 ## [0.1.3] – 2023-01-16
+
 ### Added
 - Tariff Zones added and mapped with stop places. These include the
   defined geographic polygon as GML posList.